requirements.txt
Django
djangorestframework
django-filter
reportlab
